Refactoring the package location for the TableCreatorCallback from DynamoDB to leases.
