+++ amazon-kinesis-client/src/main/java/software/amazon/kinesis/leases/LeaseRefresher.java
@@ -205,7 +205,7 @@ public interface LeaseRefresher {
 
     /**
      * Ability to perform actions of the lease table post creation. Needs
-     * {@link software.amazon.kinesis.leases.dynamodb.TableCreatorCallback} to be implemented and configured in the
+     * {@link TableCreatorCallback} to be implemented and configured in the
      * {@link LeaseManagementConfig}.
      */
     default void performPostTableCreationAction() {
